pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "icinst"
version = "0.1.0"
description = ""
requires-python = ">=3.10"
dependencies = [
    "pyslang",
    "ruamel.yaml",
    "networkx",
    "rich>=14.3.3",
]

[project.scripts]
icinst = "icinst.cli:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-mypy"
]
docs = [
    "sphinx",
    "myst-parser",
    "furo",
    "sphinx-argparse",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/icinst"]

[tool.pylint.typecheck]
# pyslang exposes its API via a compiled extension; suppress false no-member errors
generated-members = ["pyslang.*"]

[tool.pylint.format]
max-line-length = 160

[tool.pylint."messages control"]
disable = ["logging-fstring-interpolation"]

[dependency-groups]
dev = [
    "pylint>=4.0.5",
    "pytest-pylint>=0.21.0",
]
